Added deploy_group validation to all paasta instances, not just tron
